+++ app/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { ThemeProvider } from "@/components/contexts/theme-provider";
-import { Navbar } from "@/components/navbar";
+import { Navbar } from "@/components/docs/navbar";
 import { GeistSans } from "geist/font/sans";
 import { GeistMono } from "geist/font/mono";
-import { Footer } from "@/components/footer";
+import { Footer } from "@/components/docs/footer";
 import docuConfig from "@/docu.json";
 import { Toaster } from "@/components/ui/sonner";
 import "@/styles/globals.css";
@@ -45,7 +45,7 @@ export function getMetadata({
   image?: string;
 }): Metadata {
   const ogImage = image ? new URL(`/images/${image}`, meta.baseURL).toString() : undefined;
-  
+
   return {
     ...defaultMetadata,
     title: title ? `${title}` : defaultMetadata.title,