+++ components/docs/navbar.tsx
@@ -1,8 +1,6 @@
-import { ModeToggle } from "@/components/theme-toggle";
 import { ArrowUpRight } from "lucide-react";
 import Link from "next/link";
 import Image from "next/image";
-import { buttonVariants } from "./ui/button";
 import Search from "./search";
 import Anchor from "./anchor";
 import { SheetLeftbar } from "./leftbar";
@@ -10,43 +8,24 @@ import { SheetClose } from "@/components/ui/sheet";
 import docuConfig from "@/docu.json"; // Import JSON
 
 export function Navbar() {
-  const { social } = docuConfig; // Extract navbar and social from JSON
 
   return (
-    <nav className="w-full border-b h-16 sticky top-0 z-50 bg-background">
+    <nav className="sticky top-0 z-50 w-full h-16 border-b bg-background">
       <div className="sm:container mx-auto w-[95vw] h-full flex items-center justify-between md:gap-2">
         <div className="flex items-center gap-5">
           <SheetLeftbar />
           <div className="flex items-center gap-6">
-            <div className="sm:flex hidden">
+            <div className="hidden sm:flex">
               <Logo />
             </div>
-            <div className="lg:flex hidden items-center gap-4 text-sm font-medium text-muted-foreground">
+            <div className="items-center hidden gap-4 text-sm font-medium lg:flex text-muted-foreground">
               <NavMenu />
             </div>
           </div>
         </div>
 
-        <div className="flex items-center gap-3">
-          <div className="flex items-center gap-2">
+        <div className="flex items-center gap-2">
             <Search />
-            <div className="flex ml-2.5 sm:ml-0 gap-2">
-              {social.map((item) => {
-                const Icon = require("lucide-react")[item.iconName]; // Dynamically load icon
-                return (
-                  <Link
-                    key={item.name}
-                    href={item.url}
-                    target="_blank"
-                    className={buttonVariants({ variant: "ghost", size: "icon" })}
-                  >
-                    <Icon className="h-[1.1rem] w-[1.1rem]" />
-                  </Link>
-                );
-              })}
-              <ModeToggle />
-            </div>
-          </div>
         </div>
       </div>
     </nav>
@@ -59,7 +38,7 @@ export function Logo() {
   return (
     <Link href="/" className="flex items-center gap-2.5">
       <Image src={navbar.logo.src} alt={navbar.logo.alt} width="24" height="24" />
-      <h2 className="text-md font-bold font-code">{navbar.logoText}</h2>
+      <h2 className="font-bold font-code text-md">{navbar.logoText}</h2>
     </Link>
   );
 }
@@ -69,12 +48,12 @@ export function NavMenu({ isSheet = false }) {
 
   return (
     <>
-      {navbar.menu.map((item) => {
+      {navbar?.menu?.map((item) => {
         const isExternal = item.href.startsWith("http");
 
         const Comp = (
           <Anchor
-            key={item.title + item.href}
+            key={`${item.title}-${item.href}`}
             activeClassName="!text-primary md:font-semibold font-medium"
             absolute
             className="flex items-center gap-1 dark:text-stone-300/85 text-stone-800"
@@ -83,7 +62,7 @@ export function NavMenu({ isSheet = false }) {
             rel={isExternal ? "noopener noreferrer" : undefined}
           >
             {item.title}
-            {isExternal && <ArrowUpRight className="h-4 w-4 text-muted-foreground" />}
+            {isExternal && <ArrowUpRight className="w-4 h-4 text-muted-foreground" />}
           </Anchor>
         );
         return isSheet ? (